code-structure: extract bookings context from apartments
