+++ lib/ukio/apartments/bookings/bookings.ex
@@ -1,0 +1,103 @@
+defmodule Ukio.Apartments.Bookings do
+  @moduledoc """
+  The Bookings context.
+  """
+  import Ecto.Query, warn: false
+  alias Ukio.Repo
+
+  alias Ukio.Apartments.Bookings.Booking
+
+  @doc """
+  Returns the list of bookings.
+
+  ## Examples
+
+      iex> list_bookings()
+      [%Booking{}, ...]
+
+  """
+  def list_bookings do
+    Repo.all(Booking)
+  end
+
+  @doc """
+  Gets a single booking.
+
+  Raises `Ecto.NoResultsError` if the Booking does not exist.
+
+  ## Examples
+
+      iex> get_booking!(123)
+      %Booking{}
+
+      iex> get_booking!(456)
+      ** (Ecto.NoResultsError)
+
+  """
+  def get_booking!(id), do: Repo.get!(Booking, id)
+
+  @doc """
+  Creates a booking.
+
+  ## Examples
+
+      iex> create_booking(%{field: value})
+      {:ok, %Booking{}}
+
+      iex> create_booking(%{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def create_booking(attrs \\ %{}) do
+    %Booking{}
+    |> Booking.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  @doc """
+  Updates a booking.
+
+  ## Examples
+
+      iex> update_booking(booking, %{field: new_value})
+      {:ok, %Booking{}}
+
+      iex> update_booking(booking, %{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def update_booking(%Booking{} = booking, attrs) do
+    booking
+    |> Booking.changeset(attrs)
+    |> Repo.update()
+  end
+
+  @doc """
+  Deletes a booking.
+
+  ## Examples
+
+      iex> delete_booking(booking)
+      {:ok, %Booking{}}
+
+      iex> delete_booking(booking)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_booking(%Booking{} = booking) do
+    Repo.delete(booking)
+  end
+
+  @doc """
+  Returns an `%Ecto.Changeset{}` for tracking booking changes.
+
+  ## Examples
+
+      iex> change_booking(booking)
+      %Ecto.Changeset{data: %Booking{}}
+
+  """
+  def change_booking(%Booking{} = booking, attrs \\ %{}) do
+    Booking.changeset(booking, attrs)
+  end
+end
