+++ lib/ukio_web/controllers/booking_controller.ex
@@ -1,9 +1,9 @@
 defmodule UkioWeb.BookingController do
   use UkioWeb, :controller
 
-  alias Ukio.Apartments
-  alias Ukio.Apartments.Booking
-  alias Ukio.Bookings.Handlers.BookingCreator
+  alias Ukio.Apartments.Bookings
+  alias Ukio.Apartments.Bookings.Booking
+  alias Ukio.Apartments.Bookings.Handlers.BookingCreator
 
   action_fallback UkioWeb.FallbackController
 
@@ -27,7 +27,7 @@ defmodule UkioWeb.BookingController do
   end
 
   def show(conn, %{"id" => id}) do
-    booking = Apartments.get_booking!(id)
+    booking = Bookings.get_booking!(id)
     render(conn, :show, booking: booking)
   end
 end
